fix: Keep factory data hooks under the line budget
Adding notification hooks to useFactoryData.ts pushed the file
over the ESLint max-lines budget and failed CI.
