+++ web_src/src/pages/factories/pages/settings/FactorySettingsNotificationsPage.tsx
@@ -3,7 +3,8 @@ import { Checkbox } from "@/components/ui/checkbox";
 import { Label } from "@/components/ui/label";
 import { LoadingButton } from "@/components/ui/loading-button";
 import { cn } from "@/lib/utils";
-import { useFactories, useNotificationSettings, useUpdateNotificationSettings } from "@/hooks/useFactoryData";
+import { useFactories } from "@/hooks/useFactoryData";
+import { useNotificationSettings, useUpdateNotificationSettings } from "@/hooks/useNotificationSettings";
 import { getApiErrorMessage } from "@/lib/errors";
 import { showErrorToast, showSuccessToast } from "@/lib/toast";
 import { useEffect, useState } from "react";
